python/pyproject.toml
[tool.black]
line-length = 90
target-version = ['py310', 'py311', 'py312', 'py313']
skip-string-normalization = false

[tool.mypy]
strict = true

[[tool.mypy.overrides]]
module = 'TexSoup.*'
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = 'pylatexenc.*'
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = 'stop_words.*'
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = 'citeproc.*'
ignore_missing_imports = true

[tool.pytest.ini_options]
markers = [
    "integration: marks tests on the full acl-org/acl-anthology repo"
]
testpaths = ["tests"]
filterwarnings = [
    "error",
    "ignore:invalid escape sequence:SyntaxWarning"
]

[tool.ruff]
target-version = 'py310'

[tool.ruff.lint]
ignore = ['E501']  # "Line too long" is black's job

[project]
name = "acl-anthology"
packages = [
    { include = "acl_anthology" },
]
version = "0.5.3"
description = "A library for accessing the ACL Anthology"
authors = [
    { name = "Marcel Bollmann", email = "marcel@bollmann.me" },
]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/acl-org/acl-anthology"
documentation = "https://acl-anthology.readthedocs.io/"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Interface Engine/Protocol Translator",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.10, !=3.11.0, <3.14"
dependencies = [
    "docopt (>=0.6.2)",
    "lxml (~=5.3.1)",
    "PyYAML (~=6.0)",
    "app-paths (~=0.0.7)",
    "diskcache (>=5.6.1)",
    "rich (>=13.3.5)",
    "attrs (~=24.1)",
    "texsoup (>=0.3.1, <0.4.0)",
    "omegaconf (~=2.3.0)",
    "python-slugify[unidecode] (~=8.0.1)",
    "numpy (>=1.26.0)",
    "scipy (>=1.6.0)",
    "platformdirs (~=3.11.0)",
    "gitpython (>=3.1.37)",
    "rnc2rng (>=2.6.6)",
    "citeproc-py (~=0.6.0)",
    "langcodes[data] (>=3.5.0)",
    "pylatexenc (~=2.10)",
    "stop-words (>=2018.7.23)",
    "typing-extensions (>=4.6.0) ; python_version < '3.11'",
]

[project.scripts]
acl-anthology-editor = 'acl_anthology.tui.editor:main'

[tool.poetry.group.dev.dependencies]
black = ">=25.1"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
ruff = "^0.3.4"
pre-commit = "^3.3.2"
mkdocs = "^1.4.3"
mkdocstrings-python = "^1.12.0"
mkdocs-material = "^9.1.15"
mypy = "^1.3.0"
types-python-slugify = "^8.0.0.3"
types-pyyaml = "^6.0.12.11"
types-lxml = "^2023.3.28"
pymdown-extensions = "^10.13"
pytest-datadir = "^1.6.1"
richbench = {git = "https://github.com/tonybaloney/rich-bench.git"}

[tool.poetry.group.tui.dependencies]
textual = "^3.2.0"
textual-dev = "^1.7.0"
thefuzz = "^0.22.1"
dawg2 = "^0.13.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
